pyproject.toml
[project]
name = "interpretable-flow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "datasets>=3.6.0",
    "einops>=0.8.1",
    "flash-attn>=2.7.4.post1",
    "huggingface>=0.0.1",
    "huggingface-hub[hf-transfer]>=0.30.2",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "mauve-text>=0.4.0",
    "omegaconf>=2.3.0",
    "pre-commit>=4.2.0",
    "pytorch-lightning>=2.5.1",
    "ruff>=0.11.10",
    "sentencepiece>=0.2.0",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.52.3",
    "wandb>=0.20.1",
    "zstandard>=0.23.0",
    "zstd>=1.5.7.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
]
